requirements.txt
Django==4.1.4
django-cart==1.0.4
djangorestframework==3.14.0
python-decouple==3.6
requests==2.31.0
sqlparse==0.4.4
stripe==5.0.0
